First working implementation

+++ es/uca/TextAdventures/Player/Monster.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 public class Monster extends Enemy {
 
-    public Monster(String name, int id, int healthPoints, Set<Item> inventory, int baseDamage, int type) throws Enemy.TypeNotFoundException{
+    public Monster(String name, int id, double healthPoints, Set<Item> inventory, int baseDamage, int type) throws Enemy.TypeNotFoundException{
         super(name, id, healthPoints, inventory, baseDamage, type);
     }
